requirements.txt
livekit
livekit-api
cerebras-cloud-sdk
python-dotenv
httpx
numpy